backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pgvector==0.3.2
anthropic==0.39.0
openai==1.51.0
google-generativeai==0.8.2
mistralai==1.1.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
asyncpg==0.29.0
markdown==3.7
python-multipart==0.0.9
sentence-transformers==3.3.1
transformers==4.46.3
sse-starlette==1.8.2
numpy==1.26.3
torch==2.5.1
slowapi>=0.1.9
pyyaml>=6.0
